More mayoral fixes
Fixed another zeroing-out glitch in the equations for the mood bonus a colonist gets for complaining to the mayor. Colonists who complain to the mayor will now get a proper mood bonus for it. Remember, the fancier the office, and the more Empathetic the mayor, the higher the bonus.
Fixed the Sedentary trait to update the "settled in" thought properly.
Prudes who see a naked person will now properly get the negative thought for it.
Fixed some errors when pawns try to do joy activities while hanging out.
Added a letter sent when a colonist gets anxiety from a mental break.
Fixed a glitch that could cause pawns to have deep talks instead of conversations.

+++ Source/Psychology/main/Elections/LordJob_VisitMayor.cs
@@ -81,7 +81,7 @@ namespace Psychology
                         ThoughtStage complaintStage = new ThoughtStage();
                         float complaintMood = 15f * (PsycheHelper.Comp(mayor).Psyche.GetPersonalityRating(PersonalityNodeDefOf.Empathetic) - 0.33f);
                         //Base complaint mood determined by mayor's Empathetic trait
-                        complaintMood *= this.ticksInSameRoom / GenDate.TicksPerHour;
+                        complaintMood *= (float)this.ticksInSameRoom / (float)GenDate.TicksPerHour;
                         //Length of meeting also affects mood
                         complaintMood *= (complaintMood < 0f ? Mathf.Lerp(1.25f, 0.75f, PsycheHelper.Comp(constituent).Psyche.GetPersonalityRating(PersonalityNodeDefOf.Polite)) : 1f);
                         //Negative meeting thoughts (unempathetic mayors) mitigated by mayor's politeness
